requirements.txt
# Core Libraries
flask==3.0.0
streamlit==1.30.0
chromadb==0.4.15
pypdf2==3.0.1
rank_bm25==0.2.2
tqdm==4.66.1
werkzeug==3.0.1
markdown==3.8.2
# LangChain Ecosystem
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-google-genai==2.0.1
langchain-text-splitters>=0.3.0,<0.4.0


# Google AI SDK
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15

# Transformers (for fallback or hybrid search)
sentence-transformers==2.2.0
transformers==4.36.0
torch==2.1.0
tokenizers==0.15.2


